requirements.txt
natsort
numpy
opencv-python
pandas
python-box
scikit-learn
torch>=1.7.1
torchvision

albumentations
pytorchvideo
pytorch-lightning
timm
transformers
hydra-core
h5py>=3.2.0
